删除图集功能

+++ application/api/controller/Ueditor.php
@@ -4,6 +4,7 @@ namespace app\api\controller;
 use think\Controller;
 use org\UeditorUpload;
 use think\Session;
+use app\common\model\Article as ArticleModel;
 
 /**
  * Ueditor编辑器统一上传接口
@@ -14,11 +15,12 @@ class Ueditor extends Controller
 {
     protected $config;
     protected $action;
+    protected $article_model;
 
     protected function _initialize()
     {
         parent::_initialize();
-
+        $this->article_model  = new ArticleModel();
         if(!Session::get('admin_id')){
             $result = [
                 'state' => 'ERROR'
@@ -290,6 +292,8 @@ class Ueditor extends Controller
     }
     public function actionAjaxDel(){
         $file = $this->request->Param('path');
+        $id = $this->request->Param('id');
+//        dump($id);die();
         $firstchar = substr($file,0,1);
         if(in_array($firstchar, array('/','\\'))){
             $file = substr($file,1);
@@ -302,7 +306,8 @@ class Ueditor extends Controller
         try {
             $upload_path = str_replace('\\', '/', ROOT_PATH . '/');
             $file=$upload_path.$file;
-            if($file && file_exists($file)){
+            if($file && file_exists($file)&&$this->article_model->where('id', $id)
+                    ->update(['photo' => '']) !==false){
                 unlink($file);
                 $var['code'] = 1;
                 $var['state'] = 'success';
